refactor: move game modules into sub-directory
This should prevent them from being confused with identically named
files elsewhere.
