+++ game2048.koplugin/modules/history.lua
